+++ gip.core.reporthandler/Linx/LinxPrinterStatusResponse.cs
@@ -13,10 +13,10 @@
             DE	;Checksum
          */
 
-        [LinxByteMapping(Order = 1, Length = 1, DefaultValue = (short)LinxASCIControlCharacterEnum.ESC)]
+        [LinxByteMapping(Order = 1, Length = 1, DefaultValue = (byte)LinxASCIControlCharacterEnum.ESC)]
         public LinxASCIControlCharacterEnum StartCode01 { get; set; } = LinxASCIControlCharacterEnum.ESC;
 
-        [LinxByteMapping(Order = 2, Length = 1, DefaultValue = (short)LinxASCIControlCharacterEnum.ACK)]
+        [LinxByteMapping(Order = 2, Length = 1, DefaultValue = (byte)LinxASCIControlCharacterEnum.ACK)]
         public LinxASCIControlCharacterEnum StartCode02 { get; set; } = LinxASCIControlCharacterEnum.ACK;
 
         [LinxByteMapping(Order = 3, Length = 1)]
@@ -28,10 +28,10 @@
         [LinxByteMapping(Order = 5, Length = 1)]
         public LinxPrinterCommandCodeEnum CommandID { get; set; }
 
-        [LinxByteMapping(Order = 6, Length = 1, DefaultValue = (short)LinxASCIControlCharacterEnum.ESC)]
+        [LinxByteMapping(Order = 6, Length = 1, DefaultValue = (byte)LinxASCIControlCharacterEnum.ESC)]
         public LinxASCIControlCharacterEnum EndCode01 { get; set; } = LinxASCIControlCharacterEnum.ESC;
 
-        [LinxByteMapping(Order = 7, Length = 1, DefaultValue = (short)LinxASCIControlCharacterEnum.ETX)]
+        [LinxByteMapping(Order = 7, Length = 1, DefaultValue = (byte)LinxASCIControlCharacterEnum.ETX)]
         public LinxASCIControlCharacterEnum EndCode02 { get; set; } = LinxASCIControlCharacterEnum.ETX;
 
         [LinxByteMapping(Order = 8, Length = 1)]
